+++ CHANGELOG.md
@@ -1,4 +1,13 @@
+## 0.1.3 - 2017-02-24
+- Release 0.1.3 / npm@v0.1.3
+- update README.md
+- merge example1 and example2 to a single
+- extend the example2
+- add new method clone()
+
+
 ## 0.1.2 - 2017-02-23
+- Release 0.1.2 / npm@v0.1.2
 - add keywords and description
 - update examples
 - significantly update README.md